protected the routes

+++ routes/web.php
@@ -1,6 +1,7 @@
 <?php
 
 use App\Http\Controllers\ListingController;
+use App\Http\Controllers\UserController;
 use App\Models\Listing;
 use Illuminate\Support\Facades\Route;
 
@@ -18,9 +19,17 @@ use Illuminate\Support\Facades\Route;
 Route::controller(ListingController::class)->group(function () {
     Route::get('/', 'index'); // All listing
     Route::get('/listing/{listing}', 'show'); // Single listing
-    Route::get('/listings/create', 'create'); // Show create form
-    Route::post('/listings', 'store'); // Create a listing
-    Route::get('/listings/{listing}/edit', 'edit'); // Show edit form
-    Route::put('/listings/{listing}', 'update'); // Update listing
-    Route::delete('/listings/{listing}', 'destroy'); // Delete listing
+    Route::get('/listings/create', 'create')->middleware('auth'); // Show create form
+    Route::post('/listings', 'store')->middleware('auth'); // Create a listing
+    Route::get('/listings/{listing}/edit', 'edit')->middleware('auth'); // Show edit form
+    Route::put('/listings/{listing}', 'update')->middleware('auth'); // Update listing
+    Route::delete('/listings/{listing}', 'destroy')->middleware('auth'); // Delete listing
+});
+
+Route::controller(UserController::class)->group(function () {
+    Route::post('/users', 'store'); // Create a new user
+    Route::get('/register', 'register')->middleware('guest'); // Show registration page
+    Route::get('/login', 'login')->name('login')->middleware('guest'); // Show login page
+    Route::post('/logout', 'logout')->middleware('auth'); // Log user out
+    Route::post('/users/authenticate', 'authenticate'); // Log user in
 });
